requirements.txt
asgiref==3.8.1
axios==0.4.0
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Django==5.1.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
flake8==7.1.1
fonttools==4.54.1
geographiclib==2.0
geopy==2.4.1
idna==3.10
joblib==1.4.2
kiwisolver==1.4.7
lxml==5.3.0
markdown-it-py==3.0.0
matplotlib==3.9.2
mccabe==0.7.0
mdurl==0.1.2
gunicorn==21.2.0
mypy-extensions==1.0.0

nltk==3.9.1
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.18.0
pyodbc==5.1.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2024.9.11
requests==2.32.3
rich==13.9.4
seaborn==0.13.2
six==1.16.0
sqlparse==0.5.1
textblob==0.18.0.post0
tqdm==4.66.5
tzdata==2024.1
urllib3==2.2.3
